+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "javascript.validate.enable": false
+}
