+++ react_frontend/src/index.html
@@ -6,6 +6,6 @@
     </head>
     <body>
         <div id="app"></div>
-        <script type="text/javascript" src="dist/bundle.js"></script>
+        <script type="text/javascript" src="bundle.js"></script>
     </body>
 </html>
