+++ react_frontend/src/pages/App.js
@@ -1,13 +1,7 @@
 // @flow
 
 import React from 'react';
-import { graphql, QueryRenderer } from 'react-relay';
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-
-import environment from './../relay/environment';
-import StudentListItem from './../components/studentListItem';
-import Button from './../components/button';
-import addStudent from './../relay/mutations/addStudent';
 
 import StudentPage from './StudentPage';
 import TeacherPage from './TeacherPage';
@@ -40,37 +34,12 @@ class App extends React.Component<Props>{
                         </ul>
 
                         <hr />
-
                         <Route exact path="/" component={ HomePage } /> 
                         <Route path="/student" component={StudentPage} />
                         <Route path="/teacher" component={TeacherPage} />
                         <Route path="/admin" component={AdminPage} />
                     </div>
                 </Router>
-                <QueryRenderer
-                    environment={environment}
-                    query={graphql`
-                        query App_Query{
-                            students {
-                                ...studentListItem_student
-                            }
-                        }   
-                    `}
-                    variables={{}}
-                    render={({ props }) => {
-                        if (!props) {
-                            return (
-                                <div>Cant get query</div>
-                            );
-                        }
-                        return (
-                            <div>
-                                {props.students.map(s => <StudentListItem student={s} />)}
-                                {/* <Button onClick={() => addStudent(environment, 'Pranay')}>Add Pranay</Button> */}
-                            </div>
-                        );
-                    }}
-                />
             </div>
         );
     }
