+++ web_backend/src/models/student.js
@@ -5,7 +5,6 @@ const Schema = mongoose.Schema;
 const StudentSchema = new Schema({
   name: { type: String },
   email: { type: String },
-  teacherID: { type: mongoose.SchemaTypes.ObjectId, ref: 'Teacher' },
   grades: [{
     lesson: {type: String},
     score: {type: Number}
