--- web_backend/src/schema/query.js
@@ -1,40 +1,0 @@
-import { GraphQLObjectType, GraphQLList} from 'graphql';
-
-import { nodeField } from '../core'
-import Student from '../../models/student';
-import StudentType from '../types/StudentType';
-import AdminType from '../types/AdminType'
-import Admin from '../../models/admin';
-import TeacherType from '../types/TeacherType'
-import Teacher from '../../models/teacher';
-
-
-const Query = new GraphQLObjectType({
-  name: 'Query',
-  description: 'Your Root Query',
-  fields() {
-    return {
-      node: nodeField,
-      students: {
-        type: new GraphQLList(StudentType),
-        resolve() { 
-          return Student.find()
-        }
-      }, 
-      teacher: {
-        type: new GraphQLList(TeacherType),
-        resolve() {
-          return Teacher.find()
-        }
-      },
-      admin: {
-        type: new GraphQLList(AdminType),
-        resolve() { 
-          return Admin.find()
-        }
-      },
-    };
-  },
-});
-
-export default Query;
